requirements.txt
colorama
mysql.connector
pypresence
termcolor